requirement.txt
python 3.8.5
pillow
strip
django-crispy-forms
